requirements.txt
streamlit
python-docx==0.8.11
requests